Add files via upload
Write a piece of code that creates variable of each of these common data types (where possible) and follows the naming conventions:
• int
• string
• floating-point number
• boolean
• array/list
• dictionary (sometimes called a hash or a map, depending on your language)

In your code, experiment with doing different things with the data types: can you add ints and floats. Is the resulting variable an int (narrowing conversion) or a float (widening conversion)? Can you put different data types in the same array or list? Can one data type be converted to another (int to float, string to int, etc)?